renaming K8sStatusIcon to PipelineComponentStatusIcon
